conf: changed default value of (ios|android).keepalive_timeout.
