+++ gaurun/client_test.go
@@ -1,0 +1,13 @@
+package gaurun
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestKeepAliveInterval(t *testing.T) {
+	assert.Equal(t, 30, keepAliveInterval(90))
+	assert.Equal(t, 30, keepAliveInterval(30))
+	assert.Equal(t, 25, keepAliveInterval(25))
+	assert.Equal(t, 30, keepAliveInterval(50))
+}
